% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{geos_bbox}
\alias{geos_bbox}
\title{Get the bounding box of a geometry}
\usage{
geos_bbox(geom)
}
\arguments{
\item{geom}{A \link[geos:as_geos_geometry]{GEOS geometry vector}}
}
\value{
A \link[geos:as_geos_geometry]{GEOS geometry vector}
}
\description{
Get the bounding box of a geometry
}
