% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{geos_crop}
\alias{geos_crop}
\alias{geos_clip}
\title{Clip geometries to a polygon}
\usage{
geos_crop(input, overlay)

geos_clip(input, overlay)
}
\arguments{
\item{input}{A \link[geos:as_geos_geometry]{GEOS geometry vector}}

\item{overlay}{A polygon \link[geos:as_geos_geometry]{GEOS geometry vector}}
}
\value{
A \link[geos:as_geos_geometry]{GEOS geometry vector}
}
\description{
This algorithm clips a vector \code{input} layer using the features of a polygon
\code{overlay} layer.
Only the parts of the features in the input geom that fall within the
polygons of the Overlay layer will be added to the resulting layer.
}
