% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{geos_make_grid}
\alias{geos_make_grid}
\title{Create a grid of polygons over a geometry's bounding box}
\usage{
geos_make_grid(geom, nx = 5, ny = 5)
}
\arguments{
\item{geom}{A \link[geos:as_geos_geometry]{GEOS geometry vector}}

\item{nx}{Number of grid cells in x direction.}

\item{ny}{Number of grid cells in y direction.}
}
\value{
A polygon \link[geos:as_geos_geometry]{GEOS geometry vector}
}
\description{
Create a grid of polygons over a geometry's bounding box
}
\examples{
library(geos)

pts <- data.frame(
  x = runif(500, 0, 100000),
  y = runif(500, 0, 100000)
)
pts_geom <- geos::as_geos_geometry(
  wk::xy(pts$x, pts$y)
)

grid <- geos_make_grid(pts_geom)

plot(grid)
plot(pts_geom, add = TRUE)

}
