% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_along_geometry.R
\name{geos_points_along_geometry}
\alias{geos_points_along_geometry}
\title{Sample points along geometries}
\usage{
geos_points_along_geometry(geom, distance, start_offset = 0, end_offset = 0)
}
\arguments{
\item{geom}{A \link[geos:as_geos_geometry]{GEOS geometry vector}.}

\item{distance}{Distance between sampled points.}

\item{start_offset}{Distance to skip from the start of each part.}

\item{end_offset}{Distance to skip from the end of each part.}
}
\value{
A \link[geos:as_geos_geometry]{GEOS geometry vector} of multipoints.
}
\description{
Samples points along supported geometry parts at a fixed distance.
}
