% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_within.R
\name{geos_standalone_points}
\alias{geos_standalone_points}
\title{Find standalone points with no neighbors within a distance}
\usage{
geos_standalone_points(pts, maxdist)
}
\arguments{
\item{pts}{\link[geos:as_geos_geometry]{GEOS geometry vector} of points.}

\item{maxdist}{Maximum distance to consider neighbors.}
}
\value{
\link[geos:as_geos_geometry]{GEOS geometry vector}
}
\description{
Inspired by: \url{https://lbusettspatialr.blogspot.com/2018/02/speeding-up-spatial-analyses-by.html}
}
\examples{
maxdist <- 500
pts <- data.frame(
  x = runif(5000, 0, 100000),
  y = runif(5000, 0, 100000)
)
pts_geom <- geos::as_geos_geometry(
  wk::xy(pts$x, pts$y)
)
solo_pts <- geos_standalone_points(pts_geom, maxdist)

}
