% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instersect.R
\name{geos_str_intersection}
\alias{geos_str_intersection}
\title{Find geometries intersecting a polygon using a spatial index}
\usage{
geos_str_intersection(geom, polygon)
}
\arguments{
\item{geom}{A \link[geos:as_geos_geometry]{GEOS geometry vector}}

\item{polygon}{A polygon \link[geos:as_geos_geometry]{GEOS geometry vector}}
}
\value{
A \link[geos:as_geos_geometry]{GEOS geometry vector}
}
\description{
Find geometries intersecting a polygon using a spatial index
}
