% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{geos_bbox}
\alias{geos_bbox}
\title{Get the bounding box of a geometry}
\usage{
geos_bbox(geom)
}
\arguments{
\item{geom}{A \link[geos:as_geos_geometry]{GEOS geometry vector}}
}
\value{
A polygon \link[geos:as_geos_geometry]{GEOS geometry vector}
}
\description{
Returns the bounding box of all geometries in the vector.
For single geometries, the behavior of \code{geos_bbox} is
identical to \link[geos:geos_centroid]{geos_envelope}.
}
\examples{
x <- geos::as_geos_geometry(c("POINT (1 2)", "POINT (3 4)"))
geos_bbox(x)

}
