% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{geos_crop}
\alias{geos_crop}
\title{Crop geometries to a polygon}
\usage{
geos_crop(geom, polygon)
}
\arguments{
\item{geom}{A \link[geos:as_geos_geometry]{GEOS geometry vector}}

\item{polygon}{A polygon \link[geos:as_geos_geometry]{GEOS geometry vector}}
}
\value{
A \link[geos:as_geos_geometry]{GEOS geometry vector}
}
\description{
Crops a ... using a spatial index
}
